messenger/requirements.txt
python-dotenv==1.0.1
paho-mqtt==1.6.1
psutil==5.9.6
uvicorn==0.34.3
fastapi==0.115.12